Sixth import of some freebsd 13-stable features from Oct 28

+++ usr.bin/ctags/ctags.c
@@ -43,9 +43,11 @@ static char sccsid[] = "@(#)ctags.c	8.4 (Berkeley) 2/7/95";
 
 #include <sys/cdefs.h>
 #include <sys/types.h>
+#include <sys/stat.h>
 #include <sys/wait.h>
 
 #include <err.h>
+#include <errno.h>
 #include <limits.h>
 #include <locale.h>
 #include <regex.h>
@@ -142,6 +144,9 @@ main(int argc, char **argv)
 	if (!argc)
 		usage();
 
+	if (strcmp(outfile, "-") == 0)
+		outfile = "/dev/stdout";
+
 	if (!xflag)
 		setlocale(LC_COLLATE, "C");
 
@@ -163,11 +168,23 @@ main(int argc, char **argv)
 			put_entries(head);
 		else {
 			if (uflag) {
+				struct stat sb;
 				FILE *oldf;
 				regex_t *regx;
 
-				if ((oldf = fopen(outfile, "r")) == NULL)
+				if ((oldf = fopen(outfile, "r")) == NULL) {
+					if (errno == ENOENT) {
+						uflag = 0;
+						goto udone;
+					}
 					err(1, "opening %s", outfile);
+				}
+				if (fstat(fileno(oldf), &sb) != 0 ||
+				    !S_ISREG(sb.st_mode)) {
+					fclose(oldf);
+					uflag = 0;
+					goto udone;
+				}
 				if (unlink(outfile))
 					err(1, "unlinking %s", outfile);
 				if ((outf = fopen(outfile, "w")) == NULL)
@@ -197,6 +214,7 @@ nextline:
 				fclose(outf);
 				++aflag;
 			}
+udone:
 			if (!(outf = fopen(outfile, aflag ? "a" : "w")))
 				err(1, "%s", outfile);
 			put_entries(head);
